+++ src/app/(main)/(settings)/profileSettings/actions.tsx
@@ -1,0 +1,45 @@
+'use server';
+
+import { fetchWithAuth } from "@/services/fetchWithAuth";
+import { Profile } from "@/models/Profile";
+
+const symfonyUrl = process.env.SYMFONY_URL;
+
+export async function fetchUserProfileData(profileId: string): Promise<Profile> {
+    try {
+        const response = await fetchWithAuth(`${symfonyUrl}/api/profile/${profileId}`, {
+            method: 'GET',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+        });
+
+        if (!response || !response.data) {
+            throw new Error('Invalid response from the server');
+        }
+
+        return response.data as Profile;
+    } catch (error) {
+        throw new Error('Failed to fetch user profile data');
+    }
+}
+
+export async function updateUserProfileData(profileId: string, profileData: Partial<Profile>): Promise<Profile> {
+    try {
+        const response = await fetchWithAuth(`${symfonyUrl}/api/profile/${profileId}/edit`, {
+            method: 'PUT',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify(profileData),
+        });
+
+        if (!response || !response.data) {
+            throw new Error('Invalid response from the server');
+        }
+
+        return response.data as Profile;
+    } catch (error) {
+        throw new Error('Failed to update user profile data');
+    }
+}